Working

+++ include/display.h
@@ -1,3 +1,3 @@
 void initDisplay();
 void startupDisplay();
-void updateDisplay(int8_t home, int8_t guest, int8_t time);
+void updateDisplay(int home, int guest, int time);
